+++ src/test/scala/edu/ie3/simona/agent/grid/ThermalGridIT.scala
@@ -35,11 +35,12 @@ import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.Data.SecondaryData.WeatherData
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.simona.service.primary.PrimaryServiceProxy
-import edu.ie3.simona.service.weather.WeatherService
-import edu.ie3.simona.service.weather.WeatherService.Coordinate
+import edu.ie3.simona.service.weather.WeatherService.WeatherRegistrationData
+import edu.ie3.simona.service.weather.{WeatherDataType, WeatherService}
 import edu.ie3.simona.test.common.input.EmInputTestData
 import edu.ie3.simona.test.common.{DefaultTestData, TestSpawnerTyped}
 import edu.ie3.simona.test.matchers.QuantityMatchers
+import edu.ie3.simona.util.Coordinate
 import edu.ie3.simona.util.SimonaConstants.{INIT_SIM_TICK, PRE_INIT_TICK}
 import edu.ie3.simona.util.TickUtil.TickLong
 import edu.ie3.util.TimeUtil
@@ -161,9 +162,12 @@ class ThermalGridIT
       weatherService.expectMessage(
         SecondaryServiceRegistrationMessage(
           hpAgent,
-          Coordinate(
-            typicalHpInputModel.getNode.getGeoPosition.getY,
-            typicalHpInputModel.getNode.getGeoPosition.getX,
+          WeatherRegistrationData(
+            Coordinate(
+              typicalHpInputModel.getNode.getGeoPosition.getY,
+              typicalHpInputModel.getNode.getGeoPosition.getX,
+            ),
+            WeatherDataType.Current,
           ),
         )
       )
@@ -820,9 +824,12 @@ class ThermalGridIT
       weatherService.expectMessage(
         SecondaryServiceRegistrationMessage(
           pvAgent,
-          Coordinate(
-            pvInput.getNode.getGeoPosition.getY,
-            pvInput.getNode.getGeoPosition.getX,
+          WeatherRegistrationData(
+            Coordinate(
+              pvInput.getNode.getGeoPosition.getY,
+              pvInput.getNode.getGeoPosition.getX,
+            ),
+            WeatherDataType.Current,
           ),
         )
       )
@@ -839,9 +846,12 @@ class ThermalGridIT
       weatherService.expectMessage(
         SecondaryServiceRegistrationMessage(
           hpAgent,
-          Coordinate(
-            typicalHpInputModel.getNode.getGeoPosition.getY,
-            typicalHpInputModel.getNode.getGeoPosition.getX,
+          WeatherRegistrationData(
+            Coordinate(
+              typicalHpInputModel.getNode.getGeoPosition.getY,
+              typicalHpInputModel.getNode.getGeoPosition.getX,
+            ),
+            WeatherDataType.Current,
           ),
         )
       )